ray trajectory is still weird but i'm on it!

+++ srcs/tools/maths_utils.c
@@ -6,13 +6,13 @@
 /*   By: pfrances <pfrances@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/19 15:47:11 by pfrances          #+#    #+#             */
-/*   Updated: 2023/03/27 10:37:38 by pfrances         ###   ########.fr       */
+/*   Updated: 2023/03/27 14:38:03 by pfrances         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 
-float	deg_to_rad(int a)
+double	deg_to_rad(int a)
 {
 	return ((a * M_PI) / 180.0);
 }
@@ -26,14 +26,14 @@ int	fix_ang(int a)
 	return (a);
 }
 
-float	do_div(float num, float denum)
+double	do_div(double num, double denum)
 {
 	if (denum == 0)
 	{
 		if (num > 0)
-			return (FLT_MAX);
+			return (DBL_MAX);
 		else
-			return (-FLT_MAX);
+			return (-DBL_MAX);
 	}
 	else
 		return (num / denum);
